% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorrelation.R
\name{axis_prop}
\alias{axis_prop}
\title{Find a fixed proportion between the minimum and maximum values in a vector}
\usage{
axis_prop(vec, prop = 0.5)
}
\arguments{
\item{vec}{Numeric vector of data values}

\item{prop}{Numeric scalar of proportion to be calculated}
}
\value{
Scalar value of calculated proportion
}
\description{
This function finds the minimum and maximum values in a vector then
calculates a requested proportion of that. It is used to place text in
plots.
}
