% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{build_formula}
\alias{build_formula}
\title{Extracts model formula and fixed and random effects when present}
\usage{
build_formula(x, all = FALSE)
}
\arguments{
\item{x}{Model object}

\item{all}{Logical specifying whether to return random and fixed effects}
}
\value{
Returns list of model formula and random and fixed effects
}
\description{
Extracts model formula and fixed and random effects when present
}
\examples{
mod <- lm(mpg ~ disp, data = mtcars)
build_formula(mod)
}
