% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_version.R
\name{bump_version}
\alias{bump_version}
\title{Bump version up by one increment}
\usage{
bump_version(ver = get_current_version())
}
\arguments{
\item{ver}{Character string of current version. Default is to return output
of get_current_version().}
}
\value{
Named vector of incremented major, minor, patch, and developmental levels.
}
\description{
Bump version up by one increment
}
\examples{
bump_version("0.1.0")
}
