% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_element_min}
\alias{check_element_min}
\title{Test that vector has a minimum number of elements}
\usage{
check_element_min(x, num)
}
\arguments{
\item{x}{Vector}

\item{num}{Numeric scalar of minimum number of elements}
}
\value{
Error if minimum number is not met; otherwise, nothing.
}
\description{
Test that vector has a minimum number of elements
}
\examples{
check_element_min(1:5, 5)
\dontrun{
check_element_min(1:5, 6)
}
}
