% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_element_num}
\alias{check_element_num}
\title{Test that vector has exact number of elements}
\usage{
check_element_num(x, num)
}
\arguments{
\item{x}{Vector}

\item{num}{Numeric scalar of number of elements}
}
\value{
Error if exact number is met; otherwise, nothing.
}
\description{
Test that vector has exact number of elements
}
\examples{
check_element_num(1:5, 5)
\dontrun{
check_element_num(1:5, 4)
}
}
