% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_non_ascii.R
\name{check_non_ascii}
\alias{check_non_ascii}
\title{Checks all files for non-ASCII characters}
\usage{
check_non_ascii()
}
\value{
Returns data frame of files and lines with non-ASCII characters.
}
\description{
From <https://github.com/DavisVaughan/extrachecks/issues/15>
}
\examples{
nonascii <- check_non_ascii()
}
