% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_type}
\alias{check_type}
\title{Test if vector is of a specific type}
\usage{
check_type(x, type)
}
\arguments{
\item{x}{Vector}

\item{type}{Character string of vector type}
}
\value{
Error if type does not match; otherwise, nothing.
}
\description{
Test if vector is of a specific type
}
\examples{
check_type(1, "numeric")
\dontrun{
check_type("1", "numeric")
}
}
