% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_models}
\alias{compare_models}
\title{Compare a set of models}
\usage{
compare_models(..., reference = 1)
}
\arguments{
\item{...}{List of model objects}

\item{reference}{Numeric value signaling model object number to be used
for the reference model (denominator) for Bayes factors.}
}
\description{
Compare a set of models
}
