% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detritus-package.R
\docType{package}
\name{detritus-package}
\alias{detritus}
\alias{detritus-package}
\title{detritus: Collection of Miscellaneous Tools}
\description{
Miscellaneous collection of tools developed by me for me.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JeffreyRStevens/detritus}
  \item Report bugs at \url{https://github.com/JeffreyRStevens/detritus/issues}
}

}
\author{
\strong{Maintainer}: Jeffrey R. Stevens \email{jeffrey.r.stevens@protonmail.com} (\href{https://orcid.org/0000-0003-2375-1360}{ORCID}) [copyright holder]

}
\keyword{internal}
