% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{find_best_model}
\alias{find_best_model}
\title{Determines best fitting model}
\usage{
find_best_model(x, threshold = 3, ...)
}
\arguments{
\item{x}{Data frame output from `compare_models()`.}

\item{threshold}{Bayes factor threshold for distinguishing models.}

\item{...}{Additional arguments to pass on to internal functions. In
particular, if running find_best_model() inside a function, include
`envir = environment()` argument.}
}
\value{
Returns list with best fitting model object, model name, model formula, and fixed and random effects.
}
\description{
Determines best fitting model
}
\examples{
mod1 <- lm(mpg ~ disp, data = mtcars)
mod2 <- lm(mpg ~ disp + cyl, data = mtcars)
mod_comp <- compare_models(mod1, mod2)
find_best_model(mod_comp)
}
