% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_version.R
\name{get_current_version}
\alias{get_current_version}
\title{Get current package version}
\usage{
get_current_version()
}
\value{
Character string of the current package version
}
\description{
Get current package version
}
