% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorrelation.R
\name{ggcorrelation}
\alias{ggcorrelation}
\title{Plot correlations with statistics}
\usage{
ggcorrelation(
  df,
  x,
  y,
  xlabel = NULL,
  ylabel = NULL,
  cor_type = "pearson",
  side = "left",
  rank = FALSE
)
}
\arguments{
\item{df}{Data frame}

\item{x}{Variable for x-axis (must be in quotes)}

\item{y}{Variable for y-axis (must be in quotes)}

\item{xlabel}{Label for x-axis (must be in quotes)}

\item{ylabel}{Lebel for y-axis (must be in quotes)}

\item{cor_type}{Correlation type or method (must be "pearson", "spearman", or "kendall")}

\item{side}{Side of plot to place statistics text}

\item{rank}{Logical for whether rank should be plotted (TRUE) or not (FALSE)}
}
\value{
Returns a ggplot object, the correlation coefficient, and the Bayes factor
}
\description{
Plot correlations with statistics
}
\examples{
ggcorrelation(df = mtcars, x = "mpg", y = "disp")
ggcorrelation(df = mtcars, x = "mpg", y = "disp", side = "right")
}
