% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_version.R
\name{increment_version}
\alias{increment_version}
\title{Increment version in DESCRIPTION and CITATION and add new heading to
NEWS.md}
\usage{
increment_version(level = NULL, addnews = TRUE)
}
\arguments{
\item{level}{Character string "major", "minor", or "patch"}

\item{addnews}{Logical indicating whether to add new heading to NEWS.md
(default is TRUE).}
}
\value{
No output returned, but side effect of updating DESCRIPTION,
inst/CITATION, and (optionally) NEWS.md files.
}
\description{
Increment version in DESCRIPTION and CITATION and add new heading to
NEWS.md
}
