% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{num2words}
\alias{num2words}
\title{Converts numerals to words}
\usage{
num2words(x, capital = FALSE)
}
\arguments{
\item{x}{Numeric value}

\item{capital}{Logical for whether to capitalize first word (defaults to
FALSE)}
}
\value{
Returns character string of number converted to words.
}
\description{
Adapted from https://github.com/ateucher/useful_code/blob/master/R/numbers2words.r,
which is adapted from John Fox (http://tolstoy.newcastle.edu.au/R/help/05/04/2715.html)
In this version, I added an option for capitalization
}
\examples{
num2words(4)
num2words(4, capital = TRUE)
}
