% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{test_element_max}
\alias{test_element_max}
\title{Test that vector has a maximum number of elements}
\usage{
test_element_max(x, num)
}
\arguments{
\item{x}{Vector}

\item{num}{Numeric scalar of maximum number of elements}
}
\value{
Error if maximum number is exceeded; otherwise, nothing.
}
\description{
Test that vector has a maximum number of elements
}
