% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{test_element_min}
\alias{test_element_min}
\title{Test that vector has a minimum number of elements}
\usage{
test_element_min(x, num)
}
\arguments{
\item{x}{Vector}

\item{num}{Numeric scalar of minimum number of elements}
}
\value{
Error if minimum number is not met; otherwise, nothing.
}
\description{
Test that vector has a minimum number of elements
}
