% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{test_element_num}
\alias{test_element_num}
\title{Test that vector has exact number of elements}
\usage{
test_element_num(x, num)
}
\arguments{
\item{x}{Vector}

\item{num}{Numeric scalar of number of elements}
}
\value{
Error if exact number is met; otherwise, nothing.
}
\description{
Test that vector has exact number of elements
}
