% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_version.R
\name{update_citation_version}
\alias{update_citation_version}
\title{Update version in CITATION to match DESCRIPTION}
\usage{
update_citation_version(level = NULL)
}
\arguments{
\item{level}{Character string "major", "minor", or "patch"}
}
\value{
No output returned, but side effect of updating inst/CITATION file.
}
\description{
Update version in CITATION to match DESCRIPTION
}
