% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{compare_models}
\alias{compare_models}
\title{Generate AICc, BIC, and Bayes factors for model comparison}
\usage{
compare_models(..., reference = 1)
}
\arguments{
\item{...}{A list of model objects to compare}

\item{reference}{Character string for model to use as reference in model
comparison}
}
\value{
Returns a table with models, AICc, BIC, and Bayes factors.
}
\description{
Generate AICc, BIC, and Bayes factors for model comparison
}
\examples{
mod1 <- lm(mpg ~ disp, data = mtcars)
mod2 <- lm(mpg ~ disp + cyl, data = mtcars)
compare_models(mod1, mod2)
}
