% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{find_best_model}
\alias{find_best_model}
\title{Find the best model from a model comparison}
\usage{
find_best_model(x, threshold = 3)
}
\arguments{
\item{x}{Output data frame from `compare_models()`.}

\item{threshold}{Numeric threshold value for Bayes factors}
}
\description{
Find the best model from a model comparison
}
